some refactoring, separated Board component into Board and BoardContainer
