+++ src/components/light.js
@@ -1,10 +1,6 @@
 import React from 'react';
 
 export class Light extends React.Component {
-    constructor(props) {
-        super(props);
-    }
-
     onClick = () => {
         this.props.handleClick(this.props.row, this.props.col);
     };
@@ -14,9 +10,7 @@ export class Light extends React.Component {
             ? 'on'
             : 'off';
         return (
-            <button className={'grid-item ' + onOrOff} onClick={this.onClick}>
-                {this.props.shouldBeOn(this.props.row, this.props.col)}
-            </button>
+            <button className={'grid-item ' + onOrOff} onClick={this.onClick} />
         );
     }
 }
