+++ src/pages/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './style.css';
-import { Board } from '../components/board.js';
+import { BoardContainer } from '../containers/board-container';
 
 export default function Home() {
-    return <Board />;
+    return <BoardContainer />;
 }
